Fix: incorrect logic; mutating board
The issue was that I was updating a cell and was using the updated board
to calculate the state of the next cell. Instead I created a new grid,
and updated the state on that grid based on the state of a current cell.
Then finally assigned the new_grid to the board.